requirements.txt
vk_api
apscheduler
requests
psycopg2-binary